refactor and move workflow schema to workflow module
