+++ workflowparsers/lobster/__init__.py
@@ -17,3 +17,4 @@
 # limitations under the License.
 #
 from .parser import LobsterParser
+from .workflow import LOBSTERWorkflow
